Fix test imports by stabilizing adapter resolution
