+++ projects/04-llm-adapter/pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 asyncio_default_fixture_loop_scope=function
+pythonpath = .
